edit metho createdish

+++ src/main/java/com/riwi/MealMap/application/dtos/request/Ingredient/IngredientsOnlyWithName.java
@@ -14,5 +14,5 @@ public class IngredientsOnlyWithName {
     @NotNull
     private String name;
     private String measure;
-    private float quantity;
+    private double  quantity;
 }
